requirements.txt
asgiref==3.3.4
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
Markdown==3.3.4
numpy==1.21.0
pandas==1.2.5
psycopg2-binary==2.9.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
